Extract UpgradeTopBarView
